Move logging plugin code to module
