+++ src-tauri/Cargo.toml
@@ -41,6 +41,7 @@ tokio-util = "0.7.11"
 tauri-plugin-log = { git = "https://github.com/tauri-apps/plugins-workspace", branch = "v1" }
 time = "0.3.36"
 chrono = "0.4.38"
+fern = "0.6.2"
 
 [features]
 # this feature is used for production builds or when `devPath` points to the filesystem and the built-in dev server is disabled.
